removes aeson and mtl
