+++ bitcoin.cabal
@@ -1,6 +1,6 @@
 cabal-version: 1.12
 
--- This file has been generated from package.yaml by hpack version 0.34.4.
+-- This file has been generated from package.yaml by hpack version 0.35.0.
 --
 -- see: https://github.com/sol/hpack
 
@@ -93,7 +93,6 @@ library
       src
   build-depends:
       QuickCheck >=2.13.2
-    , aeson >=1.4.6.0
     , array >=0.5.4.0
     , base >=4.9 && <5
     , base16 >=0.3.0.1
@@ -108,7 +107,6 @@ library
     , hashable >=1.3.0.0
     , hspec >=2.7.1
     , memory >=0.15.0
-    , mtl >=2.2.2
     , murmur3 >=1.0.3
     , network >=3.1.1.1
     , safe >=0.3.18
@@ -136,6 +134,7 @@ test-suite spec
       Bitcoin.Keys.MnemonicSpec
       Bitcoin.KeysSpec
       Bitcoin.NetworkSpec
+      Bitcoin.Orphans
       Bitcoin.ScriptSpec
       Bitcoin.Transaction.PartialSpec
       Bitcoin.Transaction.TaprootSpec
@@ -166,7 +165,6 @@ test-suite spec
     , lens >=4.18.1
     , lens-aeson >=1.1
     , memory >=0.15.0
-    , mtl >=2.2.2
     , murmur3 >=1.0.3
     , network >=3.1.1.1
     , safe >=0.3.18
