+++ package.yaml
@@ -19,7 +19,6 @@ extra-source-files:
   - README.md
   - CHANGELOG.md
 dependencies:
-  - aeson >= 1.4.6.0
   - array >= 0.5.4.0
   - base >=4.9 && <5
   - base16 >= 0.3.0.1
@@ -34,7 +33,6 @@ dependencies:
   - hashable >= 1.3.0.0
   - hspec >= 2.7.1
   - memory >= 0.15.0
-  - mtl >= 2.2.2
   - murmur3 >= 1.0.3
   - network >= 3.1.1.1
   - QuickCheck >= 2.13.2
@@ -61,6 +59,7 @@ tests:
     verbatim:
       build-tool-depends: hspec-discover:hspec-discover
     dependencies:
+      - aeson >= 1.4.6.0
       - base64 ^>= 0.4
       - bitcoin
       - hspec >= 2.7.1
