+++ src/Bitcoin/Util/Arbitrary/Util.hs
@@ -14,25 +14,15 @@ module Bitcoin.Util.Arbitrary.Util (
     arbitraryMaybe,
     arbitraryNetwork,
     arbitraryUTCTime,
-    SerialBox (..),
-    JsonBox (..),
-    NetBox (..),
-    ReadBox (..),
-    testIdentity,
-    testSerial,
-    testRead,
-    testJson,
-    testNetJson,
     arbitraryNetData,
     genNetData,
+    toMap,
+    fromMap,
 ) where
 
 import Bitcoin.Constants
 import Bitcoin.Data
 import Control.Monad (forM_, (<=<))
-import qualified Data.Aeson as A
-import qualified Data.Aeson.Encoding as A
-import qualified Data.Aeson.Types as A
 import Data.ByteString (ByteString, pack)
 import Data.ByteString.Lazy (fromStrict, toStrict)
 import qualified Data.ByteString.Short as BSS
@@ -101,124 +91,6 @@ arbitraryNetwork :: Gen Network
 arbitraryNetwork = elements allNets
 
 
--- Helpers for creating Serial and JSON Identity tests
-
-data SerialBox
-    = forall a.
-        (Show a, Eq a, T.Typeable a, Serial a) =>
-      SerialBox (Gen a)
-
-
-data ReadBox
-    = forall a.
-        (Read a, Show a, Eq a, T.Typeable a) =>
-      ReadBox (Gen a)
-
-
-data JsonBox
-    = forall a.
-        (Show a, Eq a, T.Typeable a, A.ToJSON a, A.FromJSON a) =>
-      JsonBox (Gen a)
-
-
-data NetBox
-    = forall a.
-        (Show a, Eq a, T.Typeable a) =>
-      NetBox
-        ( Network -> a -> A.Value
-        , Network -> a -> A.Encoding
-        , Network -> A.Value -> A.Parser a
-        , Gen (Network, a)
-        )
-
-
-testIdentity :: [SerialBox] -> [ReadBox] -> [JsonBox] -> [NetBox] -> Spec
-testIdentity serialVals readVals jsonVals netVals = do
-    describe "Binary Encoding" $
-        forM_ serialVals $
-            \(SerialBox g) -> testSerial g
-    describe "Read/Show Encoding" $
-        forM_ readVals $
-            \(ReadBox g) -> testRead g
-    describe "Data.Aeson Encoding" $
-        forM_ jsonVals $
-            \(JsonBox g) -> testJson g
-    describe "Data.Aeson Encoding with Network" $
-        forM_ netVals $
-            \(NetBox (j, e, p, g)) -> testNetJson j e p g
-
-
--- | Generate binary identity tests
-testSerial ::
-    (Eq a, Show a, T.Typeable a, Serial a) => Gen a -> Spec
-testSerial gen =
-    prop ("Binary encoding/decoding identity for " <> name) $
-        forAll gen $ \x -> do
-            (runGetL deserialize . runPutL . serialize) x `shouldBe` x
-            (runGetL deserialize . fromStrict . runPutS . serialize) x `shouldBe` x
-            (runGetS deserialize . runPutS . serialize) x `shouldBe` Right x
-            (runGetS deserialize . toStrict . runPutL . serialize) x `shouldBe` Right x
-  where
-    name = show $ T.typeRep $ proxy gen
-    proxy :: Gen a -> Proxy a
-    proxy = const Proxy
-
-
--- | Generate Read/Show identity tests
-testRead ::
-    (Eq a, Read a, Show a, T.Typeable a) => Gen a -> Spec
-testRead gen =
-    prop ("read/show identity for " <> name) $
-        forAll gen $
-            \x -> (read . show) x `shouldBe` x
-  where
-    name = show $ T.typeRep $ proxy gen
-    proxy :: Gen a -> Proxy a
-    proxy = const Proxy
-
-
--- | Generate Data.Aeson identity tests
-testJson ::
-    (Eq a, Show a, T.Typeable a, A.ToJSON a, A.FromJSON a) => Gen a -> Spec
-testJson gen = do
-    prop ("Data.Aeson toJSON/fromJSON identity for " <> name) $
-        forAll gen (`shouldSatisfy` jsonID)
-    prop ("Data.Aeson toEncoding/fromJSON identity for " <> name) $
-        forAll gen (`shouldSatisfy` encodingID)
-  where
-    name = show $ T.typeRep $ proxy gen
-    proxy :: Gen a -> Proxy a
-    proxy = const Proxy
-    jsonID x = (A.fromJSON . A.toJSON) (toMap x) == A.Success (toMap x)
-    encodingID x =
-        (A.decode . A.encodingToLazyByteString . A.toEncoding) (toMap x)
-            == Just (toMap x)
-
-
--- | Generate Data.Aeson identity tests for type that need the @Network@
-testNetJson ::
-    (Eq a, Show a, T.Typeable a) =>
-    (Network -> a -> A.Value) ->
-    (Network -> a -> A.Encoding) ->
-    (Network -> A.Value -> A.Parser a) ->
-    Gen (Network, a) ->
-    Spec
-testNetJson j e p g = do
-    prop ("Data.Aeson toJSON/fromJSON identity (with network) for " <> name) $
-        forAll g $
-            \(net, x) -> dec net (encVal net x) `shouldBe` Just x
-    prop ("Data.Aeson toEncoding/fromJSON identity (with network) for " <> name) $
-        forAll g $
-            \(net, x) -> dec net (encEnc net x) `shouldBe` Just x
-  where
-    encVal net = A.encode . toMap . j net
-    encEnc net = A.encodingToLazyByteString . toMapE . e net
-    dec net = A.parseMaybe (p net) . fromMap <=< A.decode
-    name = show $ T.typeRep $ proxy j
-    proxy :: (Network -> a -> A.Value) -> Proxy a
-    proxy = const Proxy
-
-
 arbitraryNetData :: Arbitrary a => Gen (Network, a)
 arbitraryNetData = do
     net <- arbitraryNetwork
@@ -235,10 +107,6 @@ genNetData gen = do
 
 toMap :: a -> Map.Map String a
 toMap = Map.singleton "object"
-
-
-toMapE :: A.Encoding -> A.Encoding
-toMapE = A.pairs . A.pair "object"
 
 
 fromMap :: Map.Map String a -> a
