+++ stack.yaml
@@ -1,4 +1,4 @@
-resolver: lts-19.22
+resolver: lts-19.27
 system-ghc: true
 nix:
   packages:
